Revert "fix cppcheck warning"
This reverts commit 16c9c929a0265d462908b1c9d858ad03bfd51d00.

+++ src/sapphireos/keyvalue.c
@@ -152,7 +152,7 @@ static int8_t _kv_i8_persist_set_internal(
 
 static uint16_t _kv_u16_fixed_count( void ){
 
-    return ( (uint32_t)kv_end - (uint32_t)kv_start ) - 1;
+    return ( kv_end - kv_start ) - 1;
 }
 
 uint16_t kv_u16_count( void ){
